Work towards getting the route finder to work.
General
Removed RouteFinderInterface.py as it was not going to be useful
System.py
Added getPOS  to return the xyz position of the system
Modified getGateDistance to return distances in Astronomical Units
instead of kilometers
Main.py
Added an import for JumpsRoute
Added a dict for system names -> system ids
Added comments to setupRouteFinderto show how it will work once
finished.
Extended findRoute to actually setup a jumpsRoute and call it to produce
a route. (Actual route production not working yet)
GUI.py
Added import for webbrowser
Moved the configuration of the output area to configureOutput
Added tag configs and binds to output to allow for url hyperlinks
Implemented getRoute to call Main.py to return a route and then get
distance in AU and total jumps before outputting the route and
statistics.
Added buildDotlan to build a dotlan url for the route.