Progress made on Routes.py
Log
Main.py
Added some debug code so that I don't have to keep entering the systems
each time.
GUI.py
Commented out a line of code that was causing main.py to hang
Routes.py
Added comments to almost all the code
Renamed self.start and self.end to mainStart and mainEnd
Modified get_h to be able to run multiple different types of
calculation.
get_h for dl route type is working
get_h for j route type still WIP
Modified updateNode to call get_h for the correct route types based on
what type it gets passed
Added debugging code to getRoute to try and determin if the get_h
calculation for the j routeType is working correctly (Still WIP)
Added a reset method
Modified lyDistanceRoute to be able to pass it custom starts and ends
and to use default mainStart and end if nothing is passed.