Initial Stages of fix for Issue #4 completed
Previous commit started the fix by moving opened and closed arrays to be
local to the route calc, this commit changes the g,h,f and parent
storage in system to be arrays instead of single instances and adds
getters and setters for each.
Each index in those arrays is used by a different calc and are recycled
so there shouldn't be a huge amount of array indexs. Testing to be done
after this commit.

At this point the lyDistanceRoute is still working and outputting
results.