Added Avoidence list backend for Routes.py
Routes.py
__init__() now takes a new argument, avoidence, which is the avoidence
list and defaults to an empty list when no avoidence list is given
jumpsRoute() and jumpsRoute2() are currently running 2 different
implementations of avoidence list checking and automated tests are
running to find which one is quicker.

RoutesTests.py modified to have new tests for the avoidence backend

RandomTester has had some work done to make it able to test the new
avoidence list.