Started work on avoidence front end and main code
This took longer than expected to make as I was having problems with
tests and needed the UI open which didn't work and took me a few days to
figure out a workaround

Main.py
Started using setupRouteFinder() instead of building a route finder
directly in findRoute
Added a check to cleanup to assist with future tests but still allow the
program to fully exit when running normally.

Routes.py
Removed the first version of the avoidence list backend and implemented
the second version which proved fast for about 17000/20000 tests
Moved some code around to put the commented code at the bottem

GUI.py
Started working on the avoidence list front end. Not much to see yet as
the code is incomplete.