Fixed a bug in ErrorDailog.py and other misc stuff
AutoCompleteEntry.py
Fixed a bug that decided to revive itself. The AutocompleteEntrys will
now parent correctly, again -_-

ErrorDialog.py
Changed the error dialog to create multiple labels for multiple lines of
text like the help menu does, and to base its size off the amount of
lines

GUI.py
Added a check to getRoute that make sure if the user trys to find a
route without entering both an origin and a destiniation it just prints
to output instead of throwing an error

Main.py
Fixed on the the error messages having the slash for \n the wrong way
round.